更新 odfwebupgrade: 增加檢查伺服器上傳限制&比對版本、config加入odfweb版號、調整錯誤訊息

+++ apps/odfwebupgrade/lib/Controller/AdminController.php
@@ -53,6 +53,13 @@ class AdminController extends Controller {
 		$this->l10n = $l10n;
 	}
 
+	private function writeConfig() {
+		// 把目前使用的 odfweb 版號資訊寫入 config.php
+		$txtContent = file_get_contents(\OC::$SERVERROOT.'/version-odfweb.txt');
+		$currentVersion_odfweb = $txtContent ? trim($txtContent) : '0.1';
+		\OC::$server->getSystemConfig()->setValue('versionOdfweb', $currentVersion_odfweb);
+	}
+
 	/**
 	 * @return DataResponse
 	 */
@@ -72,21 +79,29 @@ class AdminController extends Controller {
 	 * @return DataResponse
 	 */
 	public function uploadZip(): DataResponse {
-
+		$folderTmp = '/updaterTmp-' . $this->config->getSystemValue('instanceid');
 		try {
 			// Check zip
 			$zipFile = $this->request->getUploadedFile('uploadZip');
 			if (!$zipFile) {
 				throw new \Exception('No uploaded file.');
 			}
+
+			if ($zipFile['error'] > 0) {
+				$errorMsg = $this->l10n->t('Unable to upload file(%s).', [$zipFile['error']]);
+				if ($zipFile['error'] === 1) {
+					$stmt = 'The uploaded file exceeds the maximum upload size [%s] of your server.';
+					$errorMsg .= $this->l10n->t($stmt, [ini_get('upload_max_filesize')]);
+				}
+				throw new \Exception($errorMsg);
+			}
+
 			if (mime_content_type($zipFile['tmp_name']) !== "application/zip") {
 				throw new \Exception('Please upload a zip file.');
 			}
 			if ($zipFile['size'] === 0)  {
 				throw new \Exception('File too small.');
 			}
-
-			$folderTmp = '/updaterTmp-' . $this->config->getSystemValue('instanceid');
 
 			// rm old tmp dir
 			if ($this->rootFolder->nodeExists($folderTmp)) {
@@ -111,24 +126,63 @@ class AdminController extends Controller {
 			$currentVersion = $txtContent ? trim($txtContent) : '0.1';
 
 			$zipTxtContent = file_get_contents('zip://' . \OC::$SERVERROOT . '/' . $filePath .'#odfweb/version-odfweb.txt');
-			if ($zipTxtContent) {
-				$updateZipVersion = trim($zipTxtContent);
-				$currentNum = str_replace( '.', '', $currentVersion);
-				$updateZipNum = str_replace( '.', '', $updateZipVersion);
-				if (intval($updateZipNum) < intval($currentNum)) {
-					throw new \Exception('Downgrade is unsupported.');
-				}
-			} else {
+			$updateZipVersion = trim($zipTxtContent);
+			if (!$updateZipVersion) {
 				throw new \Exception('Unable to read odfweb version.');
 			}
+			if (version_compare($currentVersion, $updateZipVersion, '>')) {
+				throw new \Exception('Downgrade is unsupported.');
+			}
+
+			// if (version_compare($currentVersion, $updateZipVersion, '=')) {
+			// 	throw new \Exception('已經是最新版本'); // 版號相同
+			// }
+
+			// 檢查 config.php odfweb 版號
+			$this->writeConfig();
+			$version = \OC::$server->getSystemConfig()->getValue('versionOdfweb', '0.1');
+			if (!$version) {
+				throw new \Exception('Unable to write odfweb version into config.');
+			}
+
 		} catch (\Exception $th) {
+
 			// delete updaterTmp/
-			$this->rootFolder->get($folderTmp)->delete();
+			if ($this->rootFolder->nodeExists($folderTmp)) {
+				$this->rootFolder->get($folderTmp)->delete();
+			}
 
 			return new DataResponse([
 				'data' => [ 'message' => $this->l10n->t($th->getMessage())],
 				'result' => false,
 			]);
+		}
+
+		// 檢查需不需要刪除 step
+		$stepFile = \OC::$SERVERROOT . '/data/updaterOdfweb-' . $this->config->getSystemValue('instanceid') . '/.step';
+		if(file_exists($stepFile)) {
+
+			try {
+				// 如果 step 已經是 12-end -> 刪除
+				$state = file_get_contents($stepFile);
+				if ($state === false) throw new \Exception('Could not read from .step');
+
+				$jsonData = json_decode($state, true);
+				if (!is_array($jsonData)) throw new \Exception('Cannot decode .step JSON data.');
+				$stepState = $jsonData['state'];
+				$stepNumber = $jsonData['step'];
+
+				if ($stepNumber === 12 && $stepState === "end") {
+					if(unlink($stepFile) === false) {
+						throw new \Exception('Could not rmdir .step');
+					}
+				}
+			} catch (\Exception $th) {
+				return new DataResponse([
+					'data' => [ 'message' => $this->l10n->t($th->getMessage())],
+					'result' => false,
+				]);
+			}
 		}
 
 		return new DataResponse([
